# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EmlakOS Türkiye API
  description: |
    Türkiye'nin gayrimenkul piyasası için merkezi "işletim sistemi" API'si.
    
    Bu API, ev arama, kiralama, satın alma, satma ve yönetme süreçlerinin 
    tamamını tek bir çatı altında toplayan platformun backend servislerini içerir.
    
    ## Servisler
    - **API Gateway** (Port 8080) - Ana API giriş noktası
    - **User Service** (Port 8081) - Kullanıcı yönetimi
    - **Listing Service** (Port 8082) - İlan yönetimi ve arama
    - **Valuation Service** (Port 8083) - AI destekli değerleme
    - **Notification Service** (Port 8084) - Bildirim sistemi
    
    ## Kimlik Doğrulama
    Tüm korumalı endpoint'ler JWT token gerektirir. Token'ı Authorization header'ında 
    `Bearer <token>` formatında gönderin.
    
    ## Rate Limiting
    API rate limiting uygulanmaktadır. Detaylar için response header'larını kontrol edin.
    
    ## Hata Kodları
    - `400` - Bad Request (Geçersiz veri)
    - `401` - Unauthorized (Kimlik doğrulama gerekli)
    - `403` - Forbidden (Yetki yok)
    - `404` - Not Found (Kaynak bulunamadı)
    - `429` - Too Many Requests (Rate limit aşıldı)
    - `500` - Internal Server Error (Sunucu hatası)
  
  version: 1.0.0
  contact:
    name: EmlakOS Türkiye Team
    email: api@emlakos-turkiye.com
    url: https://emlakos-turkiye.com
  
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Geliştirme ortamı
...
paths:
...
  /api/valuation/estimate:
    post:
      tags:
        - Valuation
      summary: Mülk değerleme
      description: AI destekli mülk değerleme tahmini yapar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValuationRequest'
      responses:
        '200':
          description: Değerleme tamamlandı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationResponse'
        '400':
          description: Geçersiz veri

  # Contracts
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ValuationRequest:
      type: object
      required:
        - property_type
        - features
        - location
      properties:
        property_type:
          type: string
          enum: [daire, villa, dükkan, ofis, arsa]
        features:
          type: object
          required:
            - sq_meters
            - room_count
            - building_age
          properties:
            sq_meters:
              type: number
              minimum: 1
            room_count:
              type: integer
              minimum: 1
            building_age:
              type: integer
              minimum: 0
            floor:
              type: integer
            total_floors:
              type: integer
            heating_type:
              type: string
            has_parking:
              type: boolean
            has_balcony:
              type: boolean
            is_furnished:
              type: boolean
        location:
          type: object
          required:
            - city
            - district
          properties:
            city:
              type: string
            district:
              type: string
            latitude:
              type: number
            longitude:
              type: number
        current_market_price:
          type: number
          minimum: 0
    ValuationResponse:
      type: object
      properties:
        estimated_price:
          type: number
          description: Tahmini fiyat (TL)
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
          description: Güven skoru
        price_range:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
        factors:
          type: object
          description: Değerleme faktörleri
        timestamp:
          type: string
          format: date-time
        model_version:
          type: string
...